(ZARAA-39) I have updated the code and worked on the other ticket that has to do with two seperate sections of the profile screen, I have made the view for the edit profile page and ensured that it doesnt create a new screen but using the Modal library to display the edit profile page. Colors needs to be worked on, also api needs to be implemented correctly in order to send the data.
