+++ package.json
@@ -47,6 +47,7 @@
     "react-native-dotenv": "^3.4.11",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-htmlview": "^0.17.0",
+    "react-native-modal": "^14.0.0-rc.1",
     "react-native-pell-rich-editor": "^1.9.0",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
